수정하기 기능, 필요한 파일 임포트

+++ Server/RESTFul/Update/UpdateWritingInfo.py
@@ -28,42 +28,28 @@ class UpdateWritingInfo(Resource):
         
         try:
             new_image_lines = json.loads(data.get('new_image_lines', '[]'))
+            deleteImages = json.loads(data.get('deleteImageNames', '[]'))
             
         except json.JSONDecodeError:
             return {'message': 'JSONDecodeError'}, 400
         
-        if not isinstance(new_image_lines, list):
+        if not isinstance(new_image_lines, list) and isinstance(deleteImages, list):
                 return {"message": "images must be a list of dictionaries and necessary key"}, 400
 
         print(f"hash: {hash}, title: {title}, contentText: {contentText}, images: {new_image_lines}")
-        
-        
-        originalImageObjs = ImageSchema().dump( Image.query.filter( Image.whichWriting == hash ).all() )
-        
-        originalImageName = [originalImageObj['name'] for originalImageObj in originalImageObjs]
-        originalImageIndex = [splitNameExt(name)[0][-1] for name in originalImageName]
         
         currentWriting = Writing.query.filter( (Writing.hash == hash),  ( Writing.author == validUserID )).first()
         
         newImages = []
         files = []
         
-        delImages = []
-        
         print( 'len(request.files)', len(request.files) )
         
         # new file
         for i, receivedFile in enumerate(request.files):
             
-            newFileName = receivedFile.filename
+            newImageName = receivedFile.filename
             newImageType = receivedFile.content_type
-            
-            newImageName, _ = splitNameExt(newFileName)
-            
-            for j in range(originalImageIndex):
-                if newImageName[-1] == originalImageIndex[j]:
-                    delImages.append(originalImageName)
-            
             
             newImages.append(
                 Image(
@@ -89,15 +75,19 @@ class UpdateWritingInfo(Resource):
             currentWriting.contentText = contentText
             currentWriting.modifyTime = datetime.now(timezone.utc)
             
-            if storedImagesDir and os.path.exists(storedImagesDir):
-                if delImages:
-                    for delImage in delImages:
-                        
-                        os.remove( os.path.abspath(delImage.path + delImage.name) )
-                        db.session.delete(delImage)
-                        
-                    db.session.commit()
-                if os.listdir(storedImagesDir) == []:
+            db.session.commit()
+            
+            if deleteImages:
+                for deleteImage in deleteImages:
+
+                    deleteImageFromDB = Image.query.filter( (Image.fileLocation == deleteImage[0]), (Image.name == deleteImage[1]) ).first()
+                    
+                    if deleteImageFromDB:
+                        db.session.delete( deleteImageFromDB )
+                        os.remove( os.path.abspath(deleteImage[0] + deleteImage[1]) )
+                        db.session.commit()
+                
+                if storedImagesDir and os.listdir(storedImagesDir) == [] and not newImages:
                     os.rmdir(storedImagesDir)
                 
             if newImages:
@@ -110,25 +100,7 @@ class UpdateWritingInfo(Resource):
                     files[i].save(path)
                     db.session.add(newImage)
             
-            db.session.commit()
-                        
-                        
-            
-            # if storedImagesDir and os.path.exists(storedImagesDir):
-            #     deletedImages = [storedImage for storedImage in storedImages]
-                
-            #     for remainImage in remainImages:
-            #         for deletedImage in deletedImages:
-            #             if remainImage['name'] == deletedImage.name:
-            #                 deletedImages.remove(deletedImage)
-
-            #     if deletedImages:
-            #         for deletedImage in deletedImages:
-            #             os.remove( os.path.abspath(storedImagesDir + deletedImage) )
-            #             db.session.delete(deletedImage)
-            #     # shutil.rmtree( os.path.abspath(storedImagesDir) )
-            #     # for storedImage in storedImages:
-            #     #     db.session.delete(storedImage)
+                db.session.commit()
             
             return {'message': 'Update writing Success'}, 200
